fix: processor was not loaded
